Algunas correcciones

+++ routes/routes.py
@@ -1,4 +1,6 @@
-from fastapi import APIRouter,Request, File, UploadFile, Depends, Form
+import json
+from pprint import pprint
+from fastapi import APIRouter,Request, File, UploadFile, Depends, Form, HTTPException, status
 import shutil
 from typing import Optional
 from pydantic import BaseModel
@@ -7,19 +9,13 @@ import random
 from fastapi.encoders import jsonable_encoder
 from fastapi.responses import FileResponse
 
+
 from conf.db import con
 from models.imagen import persona
-from schemas.imagen import Imagen
 
 routes = APIRouter()
-#Base model
-class Options (BaseModel):
-    id:Optional[int] = Form()
-    Nombre: str = Form()
-    direccion: str = Form()
-    dni: str = Form()
-    
-    
+
+
 @routes.get("/")
 async def get_all():
     return{
@@ -27,44 +23,7 @@ async def get_all():
     }
     
 
-#Accept request as data and file
-@routes.post("/uploadandacceptfile")
-async def upload_accept_file(options: Options = Depends(),data: UploadFile = File(...)):
-    data_options = options.dict()
-    result = "Uploaded Filename: {}. JSON Payload{}".format(data.filename,data_options)
-    return result
-
-
-@routes.post("/basic")
-async def post_basic_form(name :str = Form(...), apellido :str = Form(...), img: UploadFile = File(...)):
-    with open(getcwd() +"/"+ img.filename, "wb") as myfile:
-        content = await img.read()
-        myfile.write(content)
-        myfile.close()
-    return {"name": name, "apellido": apellido, "imagen": img.filename}
-
-
-''' @routes.post("/basic")
-async def post_basic_form(file: UploadFile = File(...)):
-    with open(getcwd() +"/"+ file.filename, "wb") as myfile:
-        content = await file.read()
-        myfile.write(content)
-        myfile.close()
-    return "success" '''
-
-@routes.post("/basic-persona")
-async def post_basic_form(name :str = Form(...), apellido :str = Form(...)):
-    return {"name": name, "apellido": apellido}
-
-
-
-
-
-
-
-
-''' ANDA!!! '''
-@routes.post("/file_mysql", tags=["Personas"])
+@routes.post("/persona_file/", tags=["Personas"])
 async def post_file(nombre :str = Form(...),apellido: str = Form(...),direccion: str = Form(...)
                     ,dni: str = Form(...),img: UploadFile = File(...)):
     with open(getcwd() +"/file/"+ img.filename, "wb") as myfile:
@@ -73,50 +32,101 @@ async def post_file(nombre :str = Form(...),apellido: str = Form(...),direccion:
         myfile.write(content)
         myfile.close()
         print(route)
-        id = random.randint(0,1000)
+        id = random.randint(0,100)
         new_image = {"id": id, "nombre": nombre, "apellido": apellido, "dni":dni, "direccion": direccion ,"dirImg": route}
-        result = con.execute(persona.insert().values(new_image))
+        con.execute(persona.insert().values(new_image))
+        con.commit()
         return {
             "success": True,
-            "message": "Se agrego la imagen",
+            "message": "La persona se agrego correcctamente",
         } 
         
 
-@routes.get("/file_mysql/", tags=["Personas"])
+@routes.get("/persona_file/", tags=["Personas"])
 async def get_files():
     result = con.execute(persona.select()).fetchall()
-    return result
+    pprint(result)
+    try:
+        response = []
+        for data in result:
+            img = FileResponse(data.dirImg)
+            response.append({
+                    "nombre": data.nombre,
+                    "apellido": data.apellido,
+                    "dni": data.dni,
+                    "direccion": data.direccion,
+                    "dirImg": img
+                    })
+        return response
+    except Exception as err:
+            raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT,
+                detail="Error en solicitud") from err
 
-@routes.get("/file_mysql/{id}", tags=["Personas"])
+@routes.get("/persona_file/{id}", tags=["Personas"])
 async def get_on_file(id:int):
     result = con.execute(persona.select().where(persona.c.id == id)).first()
-    img = FileResponse(result.dirImg)
-    return {"id": result.id, "name": result.nombre, "apellido": result.apellido, "dni":result.dni, "direccion": result.direccion,
-            "img": img}
+    pprint(result)
+    if result == None:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="No existe la persona ingresada"
+        )
+    try:
+        img = FileResponse(result.dirImg)
+        return {
+            "id": result.id,
+            "name": result.nombre,
+            "apellido": result.apellido,
+            "dni":result.dni,
+            "direccion": result.direccion,
+            "img": img
+            }
+    except Exception as err:
+        raise HTTPException(
+            status_code=status.HTTP_409_CONFLICT,
+            detail="Error en solicitud") from err
 
-@routes.delete("/file_mysql/{id}", tags=["Personas"])
+@routes.delete("/persona_file/{id}", tags=["Personas"])
 async def delete_file(id:int):
     result = con.execute(persona.select().where(persona.c.id == id)).first()
+    if not result:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="No existe la persona ingresada"
+        )
     remove(result.dirImg)
     con.execute(persona.delete().where(persona.c.id == id)).first()
+    con.commit()
     return "Borrado"
 
 
-@routes.put("/file_mysql/{id}", tags=["Personas"])
+@routes.put("/persona_file/{id}", tags=["Personas"])
 async def update_user(id:int, nombre :str = Form(...),apellido: str = Form(...),direccion: str = Form(...)
                     ,dni: str = Form(...),):
-
+    result = con.execute(persona.select().where(persona.c.id == id)).first()
+    if not result:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="No existe la persona ingresada"
+        )
     con.execute(persona.update().values(nombre = nombre, apellido = apellido, dni = dni, direccion = direccion)
                 .where(persona.c.id == id))
+    con.commit()
     return {
         "success": True,
         "message": "Se modifico los datos de la imagen"
     } 
 
-@routes.put("/file_mysql_img/{id}", tags=["Personas"])
+@routes.put("/persona_file_img/{id}", tags=["Personas"])
 async def update_user(id:int, img: UploadFile = File(...)):
 
     result = con.execute(persona.select().where(persona.c.id == id)).first()
+    if not result:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="No existe la persona ingresada"
+        )
     remove(result.dirImg)
 
     with open(getcwd() +"/file/"+ img.filename, "wb") as myfile:
@@ -126,27 +136,8 @@ async def update_user(id:int, img: UploadFile = File(...)):
         myfile.close()
         print(route)
     con.execute(persona.update().values(dirImg = route).where(persona.c.id == id))
+    con.commit()
     return {
         "success": True,
         "message": "Se modifico los datos de la imagen"
-    } 
-    
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
+    }
